requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
asyncpg
cloudinary
pyjwt
pydantic
passlib
